:coffin: refactor: remove dead code
