+++ src/renderer/components/refactor/display-example-projects/index.tsx
