+++ src/renderer/components/refactor/display-recent-projects/index.tsx
