+++ src/renderer/features/screens/start/components/index.ts
@@ -1,3 +1,1 @@
 export { default as ActionsBar } from './actions-bar'
-export { default as DisplayExampleProjects } from './display-example-projects'
-export { default as DisplayRecentProjects } from './display-recent-projects'
