+++ src/renderer/features/screens/workspace/components/panels/navigation-panel/index.tsx
@@ -1,6 +1,5 @@
 import { ReactNode } from 'react'
 
-import { NavigationPanelBreadcrumbs } from './breadcrumbs'
 import { NavigationPanelTabs } from './tabs'
 
 // const TestData = {
@@ -17,17 +16,6 @@ export const NavigationPanel = (): ReactNode => {
   return (
     <div className='h-[70px] w-full overflow-hidden rounded-lg border-2 border-neutral-200 bg-white dark:border-neutral-850 dark:bg-neutral-950'>
       <NavigationPanelTabs />
-      <NavigationPanelBreadcrumbs
-        crumb={{
-          key: '1',
-          project_name: 'Project Name',
-          pou_to_display: {
-            name: 'Pou Name',
-            type: ['program'],
-            language: ['ld'],
-          },
-        }}
-      />
     </div>
   )
 }
